direnv: enable nix-direnv
Trying to figure out a nice solution for having GC roots exist for
devShells so GC runs don't remove them, and one way is to let direnv
manage a profile directory for them.

Unfortunately direnv struggles in cases where a new shell is launched
as part of devShell activation, looping as it attempts to apply the
environment to each newly spawned shell.

Maybe that problem is fixed in nix-direnv, which provides its own use_*
implementations (for nix (channels) and flake).