bitwarden: enable the new SSH agent
Since for the time being it seems the Flatpak running in its own process
namespace prevents it from identifying the requesting application and
makes it crash, we'll just switch to using the Nix-packaged app for now.

Once they cut a release including bitwarden/clients#13324 we should be
good to switch back.