component/firefox: only manage search profiles with a local set
This isn't a particularly elegant solution, but it should get away from
file conflict issues templating search.mozlz4. It seems that Firefox
replaces Home Manager's link to this file with an actual file on start,
which then causes Home Manager to complain that the file already exists
and is not a link into the next store.

I don't imagine there's a better solution to this problem given that
Firefox is (sensibly) trying to prevent search engine hijacking, and my
use case is very niche.